refactor: move client logic to own component
