--- apps/core/app/[locale]/(default)/blog/[blogId]/_components/sharing-links/fragment.tsx
@@ -1,13 +1,0 @@
-import { graphql } from '~/client/graphql';
-
-export const SharingLinksFragment = graphql(`
-  fragment SharingLinksFragment on BlogPost {
-    entityId
-    thumbnailImage {
-      url: urlTemplate
-    }
-    seo {
-      pageTitle
-    }
-  }
-`);
